+++ src/QueryBuilder/QueryBuilder/QueryBuilder.js
@@ -6,7 +6,7 @@ import { Button } from '@folio/stripes/components';
 import PropTypes from 'prop-types';
 import { QueryBuilderModal } from './QueryBuilderModal';
 
-export const QueryBuilder = ({ disabled, initialValues }) => {
+export const QueryBuilder = ({ disabled, saveBtnLabel, initialValues }) => {
   const [isModalShown, setIsModalShown] = useState(false);
 
   const openModal = () => {
@@ -26,6 +26,7 @@ export const QueryBuilder = ({ disabled, initialValues }) => {
       <QueryBuilderModal
         setIsModalShown={setIsModalShown}
         isOpen={isModalShown}
+        saveBtnLabel={saveBtnLabel}
         initialValues={initialValues}
       />
     </>
@@ -34,5 +35,6 @@ export const QueryBuilder = ({ disabled, initialValues }) => {
 
 QueryBuilder.propTypes = {
   disabled: PropTypes.bool,
+  saveBtnLabel: PropTypes.string,
   initialValues: PropTypes.object,
 };